Change to use write data
